backend/requirements.txt
asgiref==3.7.2
django-cors-headers==4.3.0
django-rest-framework==0.1.0
django-sslserver==0.22
Django==4.2.7
djangorestframework==3.14.0
pip==23.2.1
PyJWT==2.8.0
pytz==2023.3.post1
setuptools==68.2.0
sqlparse==0.4.4
tzdata==2023.3
wheel==0.41.2
